Add boxed specializations of BytesBuiltin.FindNode
Prior to this commit, pytz.timezone fails with an
UnsupportedSpecializationException because it calls bytes.find with a
(boxed) Long rather than an int for the argument `start`. To fix this,
I've added new specializations that expect a Number, which should allow
this method to work when passed in boxed number types.