Fix race condition when contexts share an engine
This patch includes fixes for a category of race conditions when
concurrently running code in two or more contexts that share an engine.

The PythonNAryClinicBuiltinNodes can have multiple threads call
castWithNode concurrently on the same object. After one thread has
assigned a particular castNode, another can overwrite the array with an
empty one, causing an NPE when re-reading it from the array.  Instead,
the methods can assign directly a local variable and then save the
values in the array. With this change, a thread doesn't risk reading a
null after having computed the value.